api_yamdb/api/mixins.py
from rest_framework import mixins


class MyCustomViewSet(mixins.CreateModelMixin,
                      mixins.DestroyModelMixin,
                      mixins.ListModelMixin,
                      viewsets.GenericViewSet):
    pass